feat: added test for header and homepage
