+++ src/pages/HomePage.test.tsx
@@ -1,14 +1,19 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import HomePage from './HomePage';
+import { BrowserRouter } from 'react-router-dom';
 
 describe('Home Page', () => {
+  test('should be render home page', () => {
+    const link = render(<HomePage />, { wrapper: BrowserRouter });
+    expect(link).toBeTruthy();
+  });
   test('render search', () => {
-    render(<HomePage />);
+    render(<HomePage />, { wrapper: BrowserRouter });
     expect(screen.getByPlaceholderText(/Search card/i)).toBeInTheDocument();
   });
   test('render cards', () => {
-    render(<HomePage />);
+    render(<HomePage />, { wrapper: BrowserRouter });
     expect(screen.getByTestId('content')).toBeInTheDocument();
   });
 });
