+++ src/styles/style.css
@@ -17,13 +17,16 @@
 }
 
 .content {
+  margin: 0 auto;
+  max-width: 900px;
   display: grid;
-  grid-template-columns: repeat(3, 1fr);
+  grid-template-columns: repeat(4, 1fr);
   gap: 1rem;
   padding: 20px;
 }
 
 .header {
+  font-weight: 700;
   display: flex;
   justify-content: space-around;
   align-items: center;
@@ -55,15 +58,6 @@
   margin: 0;
 }
 
-.footer {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  width: 100%;
-  height: 50px;
-  border-top: 2px solid lightgray;
-}
-
 .card {
   display: flex;
   flex-direction: column;
@@ -86,5 +80,26 @@
 
 .search {
   margin: 20px auto 0;
-  width: 200px;
+  width: 300px;
+  height: 25px;
+  padding: 0 10px;
+  border-radius: 5px;
+}
+
+@media (max-width: 985px) {
+  .content {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+@media (max-width: 750px) {
+  .content {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media (max-width: 450px) {
+  .content {
+    grid-template-columns: repeat(1, 1fr);
+  }
 }
